+++ ffi/rodbus-ffi/Cargo.toml
@@ -24,6 +24,7 @@ num_cpus = "1"
 [build-dependencies]
 rodbus-schema = { path = "../rodbus-schema" }
 rust-oo-bindgen = { git = "https://github.com/stepfunc/oo_bindgen.git", tag = "0.3.0" }
+tracing-ffi-schema = { git = "https://github.com/stepfunc/tracing-ffi.git", tag = "0.1.0" }
 
 [features]
 default = ["serial", "tls"]
