+++ ffi/rodbus-ffi/build.rs
@@ -1,5 +1,15 @@
+use std::env;
+use std::io::Write;
+use std::path::Path;
+
 fn main() {
     println!("cargo:rerun-if-changed=build.rs");
+
+    let mut file =
+        std::fs::File::create(Path::new(&env::var_os("OUT_DIR").unwrap()).join("tracing.rs"))
+            .unwrap();
+    file.write_all(tracing_ffi_schema::get_impl_file().as_bytes())
+        .unwrap();
 
     match rodbus_schema::build_lib() {
         Ok(lib) => {
