+++ ffi/rodbus-ffi/src/tracing.rs
@@ -1,0 +1,1 @@
+include!(concat!(env!("OUT_DIR"), "/tracing.rs"));
